Move Pump model to its own file
The Pump model is currently living right in the FeederController file.
This is a bit awkward because it clutters the FeederController file.
Down the road, if the pump is used in another file, it'll also make it
hard to autoload the Pump model.

Move the Pump model to app/models.  Even though it does not use
ActiveRecord, it is definitely part of the model layer.