Make the serial port wrapper class configurable
I'd like to avoid talking to a serial port on my Mac in development
mode; it's not really hooked up to an Arduino.  So I need the serial
port to be configurable to be a stub.

Also, it's best if the Pump class knows just a bit less about the
default serial port.

Inject the configuration from the controller, because that already
knows about Rails, and it seems acceptable for it to know that a
hardware configuation exists (though not that it's necessarily
serial-specific).  Relieve the Pump model of knowing which class to
instanciate to create a serial port wrapper.