Create a SimpleSerialPort wrapper for SerialPort
I need to mock this boundary of the application.  Best practice says that
I should only mock APIs owned by the application.  Also, the interface of a
SerialPort is far too complex to faithfully mock.

Wrap this vendor library with an interface owned by the application, one
which is far simpler.