Remove serial port details from Pump model
The rest of the application shouldn't be required to know about the
details of a SerialPort when it deals with a Pump.  Also, the pump
should use our new SimpleSerialPort wrapper for SerialPort.

Switch to the narrower SimpleSerialPort interface, and stop passing
the path of the serial port from the FeederController.  This means
FeederController is now completely decoupled from serial ports, and
the Pump model is better focussed on its responsibility: controlling a
pump.