Extract feeder#dose to a Doser method object
The dose action is a bit complicated for a controller action.

Refactor it by extracting to its own object which can manage the dosing
process.  Break apart the method into helpers so that each helper method
is brief, focussed, and at a single level of abstraction.

Split the asynchronisity into a separate API entry point to facilitate
testing within a single transaction.  (Transactional fixtures are
incompatible with multithreaded tests.)