Remove Doser's dependency on a Rails environment
Doser requires a full rails environment for testing because it
creates a Dose ActiveRecord model.

Pass in the active record class as a factory object.  This flex
point allows a mock to be easily injected during testing.  This
in turn allows the spec to run without loading Rails and without
hitting the database.  It also improves code quality by reducing
coupling between the Doser and Rails.