Allow an in-progress dose to be cancelled
Once a dose is in progress, there is currently no way to modify or
stop it, short of restarting the server.  And even then, there is
no guarantee that all pumps will be left in the "off" state.  Even
if the "Starve" action is invoked, the pump could still be switched
back on by the background thread.

Allow an in-progress dose to be cancelled:

* Change the "Completed at" column on the index page to a Status column
* Add "cancel" link under Status for an in-progress dose
* Process the "cancel" action by cancelling a dose
* Add a cancelled_at column to the doses table to record the
  cancellation
* Stash away the thread when dosing and kill when cancelling