Add failed state and switch fto separate state column
We need to add a state: failed.  This allows us to show the user when a
dose has failed rather than only logging it to the application log file.

Alter the doses table to store the state and finished timestamp
separately rather than one timestamp per ending state, so that we don't
have three ending state timestamps.  Update these new columns with
timestamp and failed state when handling a dosing error.