Order doses so recent ones appear first
Newly created Doses currently get added at the bottom, but that's a bit
disorienting because they are below the fold.

Add them to the top of the list instead by reversing the sort order
of the doses.  Add a spec for this, and tolerate a few warts in the
controller setup for now to avoid errors:

* setting @controller explicitly
* rendering views