Stop initializing pump in controller constructor
In controller specs, I can't seem to stub SimpleSerialPort.new soon
enough: the before(:each) callback runs *after* the controller under
test is created.  As a result, creating the pump raises an error about
the serial port not being present.  The test somehow passes anyway,
but it breaks the final initialization steps of the example which set
@controller and setup view stubbing details.  The result is these errors
unless I add some work-arounds to the spec:

    @controller is nil: make sure you set it in your test's setup
    method.

and:

    NoMethodError: undefined method `original_path_set' for
    nil:NilClass

Create the pump lazily in a private attribute reader of the controller
and clean up the work-arounds in the spec.  We don't even need to mock
the SimpleSerialPort because it won't be used in this example.