Add unique constraint on readings sensor_id+measured_at
It wouldn't make sense to have two readings for the same sensor at the
same time, but the model doesn't protect against that.

Add a unique constraint on readings (sensor_id, measured_at).  It will
offer useful protection, and it will likely support many of the queries
that we'll do when reporting on this data.