Generate a Sensor scaffolds and remove unneeded parts
We're gonna need more machinery around sensor readings.  It would
be ideal to have it conform to Rails conventions as much as possible.

Generate the scaffold as a starting point, minus the already-existing
model and model specs, and remove bits we don't need, keeping the
controller, views, stylesheet, and route.