Chart sensor readings with NVD3 on sensors#show
We've got lots of sensor readings, and nothing conveys trends and
anomalies in a time series like a pretty graph.

Show a graph of sensor readings over time on the sensors#show page.  Use
NVD3, a library that builds charts based on D3.  Convert the show
template to slim for easy coding, but keep the chart in an ERB partial
for precise control over a little inline JS.  Keep the JS inline so we
can easily inject customized axis configurations in the future.
Configure Slim to output pretty HTML rather than space-stripped HTML to
improve debuggability of markup.