Show current status of readings on sensor index
The sensor index shows its attributes.  But we need a way to convey
the current state of the system.

Add the current sensor value and trend to the sensor index page.

It would be nice to have a more full-featured dashboard here, for
example by showing wheather a measurements is in-bounds, by showing
the trend over the last N measurements, showing the current
value as the average of the last N measurements, and/or showing charts
of all of the sensor readings.  But let's start simple.