Refix parsing of dates to work in UTC or other timezones
This spec is failing:

  1) Feeding Dosing shows past doses
     Failure/Error: page.should have_content '2013-01-01 08:00'
       expected to find text "2013-01-01 08:00" in "Sensa BioBot
            SensorsFeeder About Feed Me Total quantity Number of cycles Pause
            between cycles Dose History Total Quantity Number of Cycles Pause
            Between Cycles Status Finished 4.2 3 40.0 Completed 2013-01-01
            16:00:00 UTC Terms of Service © Sensa Bio 2012"
     # ./spec/features/pump_control_spec.rb:73:in `block (4 levels) in <top (required)>'
     # ./spec/features/pump_control_spec.rb:61:in `block (3 levels) in <top (required)>'

Travis runs specs in UTC, where this spec now passes.  But it's failing
in my local development environment, where the timezone is US/Pacific.
It seems that the current spec only works in UTC because Time.parse
always parses a date without zone as UTC, whereas the date is formatted
according to the local timezone.

Explicitly parse the date in the local timezone so that it will come
out the same when formatted in the local timezone.