Attempt to fix intermittent failing specs for dosing exceptions
This spec is still failing sometimes, but now it's because finished_at
is unset, not just out of bounds.

  1) Feeding Dosing doses
     Failure/Error: dose.finished_at.to_f.should be_within(0.3).of(DateTime.now.to_f)
       expected 0.0 to be within 0.3 of 1375018523.2004826
     # ./spec/features/pump_control_spec.rb:54:in `block (3 levels) in <top (required)>'

We're checking whether the timestamp is within +- 0.3 sec of now... but
of course it will not be in the future!  So instead wait 0.3 additional
seconds for the finished_at to be set in case it's set late.